redo to match dev_checks style
